Create FeatRedFixedTfidf.scala
Data is reduced from 5 million lines of definitions, to 5K lines of definitions, which still gives out of memory error during the training for a naive bayes classifier. Reducing the data more would make dilute the semantics of the problem; word-based feature approach will be left and letter-based feature approach will be tried.