Create LetterFeaturesNaiveBayes
Use the frequency of each letter, e.g., 'A', 'B', etc..., in the definitions of each word as a feature vector. Then, use naive Bayes classifier to classify the data. We use only the crosswords as the training data, and we get 71,4 % accuracy. If we includeiktionary data to the training data the accuracy drops to 37 %, hence we excluded the Wiktionary data.