Create LetterFeaturesDecisionTree.scala
Tries DecisionTree with letter features instead of Naive Bayes. However, it takes much longer than Naive Bayes and requires a lot of tuning, which also takes a lot of time. The parameters that needs to be tuned: maxDepth and maxBin; with maxDepth = 10 and maxBin = 16, and using half of the crosswords, it finishes in 23 minutes and gives only 7% accuracy. When we increase the maxDepth = 20 and maxBin = 32, again with the half of the crossword data, it cannot finish in 1 hour (upon when we killed the process).