Create TfidfNaiveBayes.scala
Uses Naive Bayes classifier not with the frequency of each letter in the definitions of the words, but the tf-idf value of each letter in each word across the whole crosswords data. The accuracy is 55% which is less than the result of using Naive Bayes classifier with frequency of each letter in the definitions of the words (73 %).